Moves home and browse
