+++ src/components/Navigation.js
@@ -8,17 +8,18 @@ const Navigation = () => {
         <img src="./icons/home-icon.svg" alt='home icon'/>
        </Link>
 
-      <Link id="browse-nav" to="/types">
+      <Link id="browse-nav" to="/all">
         <img src="./icons/grid-icon.svg" alt='grid icon'/>
       </Link>
 
+      <Link id="dashboard-nav" to="/types">
+      <img src="./icons/person-icon.svg" alt='person icon'/>
+      </Link>
+      
       <Link id="location-nav" to="/location">
         <img src="./icons/location-icon.svg" alt='location icon'/>
       </Link>
 
-      <Link id="dashboard-nav" to="/dashboard">
-        <img src="./icons/person-icon.svg" alt='person icon'/>
-      </Link>
 
     </nav>
   );
